Remove name mangling from find_dbus_interfaces
Double underscore in Python mangles names and sometimes is
used as a way to make things private; however, in general
this isn't considered best practice and there isn't a compelling
reason for it here.

Rather, the convention for denoting a private interface is
to use a single leading underscore.

For more information:
http://www.artima.com/weblogs/viewpost.jsp?thread=211430

Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>