pyphosphor: Add parameter to Emergency signal of SensorThresholds
When sensor value is out of range of threshold, an "Emergency"
signal will be emitted. Add a 'message' parameter to the signal,
to provide detail information on the cause of 'Emergency'.
This information will be logged in an ESEL.

This patch is part of the fix for issue:
openbmc/skeleton/issues#64

Change-Id: I501188b36718b8a87582558c349c0316e1092db5
Signed-off-by: Yi Li <adamliyi@msn.com>