Point PowerSupplyRedundancySensor to new interface
The PowerSupplyRedundancy sensor is used by the Host to know
whether the system has redundant power supply support.
The sensor is implemented under the legacy openbmc/pyphosphor
repo which does not persist it's values. This commit will
move this sensor into the official phosphor settings repo
(which does have persistency support) but also keep the legacy
interfaces in place for backwards compatibility.

The PowerSupplyRedundancySensor should look at the new
PowerSupplyRedundancy interface for its value and when
setting the PowerSupplyRedundancySensor, set the new value.

Resolves openbmc/openbmc#2833

Change-Id: I490d385efba70f26b1a3eac87074b1fa0f8cd907
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>