sensors: Remove sensor manager
This sensor manager was in charge of creating the
/org/openbmc/sensors object and initializing them, but this path
is now deprecated and the sensors have moved to other paths
specified by the ipmi sensor config.yaml file in the
openbmc/openbmc repo. This sensor manager can now be removed.

Tested: Powered on a Witherspoon to the Host OS without ipmi
giving get/set sensor errors.

Change-Id: Id9f7626d84d49c62cdb48b5d75f4247ed5813b4c
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>